+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added spellcheck dictionary: English (UK) - [#417](https://github.com/hkdb/aerion/issues/417)
 - Added spellcheck dictionary: Dutch
 - Added dot and glow dot optional accents for unread messages
+- Added window size persistence - [#217](https://github.com/hkdb/aerion/issues/217)
 - Fixed post delete ghost messages on message list
 - New app icon with full hicolor size set + scalable SVG - [#395](https://github.com/hkdb/aerion/issues/395) (Special thanks to @gianlucamazza) - [#99](https://github.com/hkdb/aerion/issues/99) 
 - Fixed thread sender and preview in message list - [#169](https://github.com/hkdb/aerion/issues/169)